requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
next==14.2.30
react==18.3.1
react-dom==18.3.1
typescript==5.0.4
tailwindcss==3.3.0
lucide-react==0.263.1
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
rapidfuzz==3.4.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 